Migrated black box to separate file
